Add app_settings.dart.
Add bottomNavigationBar to BasePage.

The height of the BottomAppBar is determined by applying a scale factor to the height of the appBar. Since the height of the appBar is required, and whilst the BasePage is being built the scaffold height is not known, a Builder widget is required in order to extract the appBar height from its parent.