Add Button class to SlidingButton.
This commit required Animation<double> to be specified everywhere. Leaving it as just Animation meant that Animation<double> got assigned to Animation<dynamic>, which generates a fail.

Note also that since Button.animation is potentially null, the ! cast operator is used in the Button class within the associated else block.